+++ test/key_provider_test.rb
@@ -8,8 +8,7 @@ class KeyProviderTest < Minitest::Test
   def test_create
     # should ideally be encrypt: 1 and no decrypt
     # https://github.com/rails/rails/issues/42388
-    encrypt = ActiveRecord::VERSION::STRING.to_f >= 7.1 ? 1 : 3
-    assert_operations encrypt: encrypt, decrypt: 1 do
+    assert_operations encrypt: 1, decrypt: 1 do
       User.create!(email: "test@example.org")
     end
   end
@@ -27,9 +26,7 @@ class KeyProviderTest < Minitest::Test
 
     # should ideally be encrypt: 1 and no decrypt
     # https://github.com/rails/rails/issues/42388
-    encrypt = ActiveRecord::VERSION::STRING.to_f >= 7.1 ? 1 : 3
-    decrypt = ActiveRecord::VERSION::STRING.to_f >= 7.1 ? 4 : 3
-    assert_operations encrypt: encrypt, decrypt: decrypt do
+    assert_operations encrypt: 1, decrypt: 4 do
       user.update!(email: "test2@example.org")
     end
   end