+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 ## 0.2.0 (unreleased)
 
-- Dropped support for Ruby < 3.1
+- Dropped support for Ruby < 3.2 and Active Record < 7.1
 
 ## 0.1.1 (2023-12-01)